requirements.txt
# Google Ads API
google-ads>=19.0.0

# Google AI/Gemini
google-generativeai>=0.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Scheduling
schedule>=1.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.62.0
PyYAML>=5.4.0
joblib>=1.1.0

# Logging and error tracking
rich>=12.0.0
wandb>=0.12.0

# Core ML / Scientific Computing
scikit-learn>=1.3.0
scipy>=1.10.1
statsmodels>=0.14.0
torch>=1.13.1
# tensorflow>=2.12.0  # Commented out due to installation issues on Python 3.13/Windows
# tensorflow-probability>=0.19.0 # Depends on tensorflow

# Reinforcement Learning
gymnasium>=0.28.1
stable-baselines3[extra]>=1.7.0
tensorboard>=2.12.0
sb3-contrib>=1.7.0
imitation>=0.3.0
ale-py>=0.8.0
autorom>=0.6.1
moviepy>=1.0.3

# Multi-armed Bandits / Bayesian
pymc>=5.0.0
arviz>=0.14.0

# Meta Learning / Hyperparameter Optimization
optuna>=3.2.0
# ray[tune]>=2.5.1  # Commented out due to installation issues on Python 3.13/Windows
scikit-optimize>=0.9.0

# Forecasting
prophet>=1.0.0
pmdarima>=2.0.3
neuralprophet>=0.6.0

# Personalization / Clustering
seaborn>=0.11.0
umap-learn>=0.5.3
hdbscan>=0.8.0

# SERP Scraping / Web Interaction
selenium>=4.10.0
beautifulsoup4>=4.10.0
webdriver-manager>=3.8.0
html5lib>=1.1
playwright>=1.35.0

# Causal Inference
causalimpact>=1.2.1
econml>=0.14.1

# Portfolio Optimization / OR-Tools
cvxpy>=1.3.1
pulp>=2.7.0
ortools>=9.6.0

# LTV Bidding and Optimization / Explainability
xgboost>=1.7.0
lightgbm>=3.3.5
shap>=0.41.0

# Graph Optimization / Visualization
networkx>=3.0
pyvis>=0.3.2

# Contextual Signals
geopy>=2.3.0
holidays>=0.21

# Trend Forecasting Specific
pytrends>=4.9.0

# Web UI / Feedback UI
streamlit>=1.28.0
flask>=2.0.0
jinja2>=3.0.0
wtforms>=3.0.0
email-validator>=1.1.0
fastapi>=0.68.0
uvicorn>=0.15.0
aiohttp>=3.8.0

# Google Cloud / API Core (Consolidated)
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
google-auth-httplib2>=0.1.0
google-api-core>=2.11.0
google-cloud-core>=2.3.2
google-cloud-storage>=2.8.0
google-crc32c>=1.5.0
google-resumable-media>=2.5.0
googleapis-common-protos>=1.59.0

# Typing and Base Utilities
typing-extensions>=4.5.0
dataclasses>=0.8
uuid>=1.30

# Testing and Development Tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.1.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0
types-requests>=2.31.0.1
types-setuptools>=67.8.0.0
types-urllib3>=1.26.25.13
pandas-stubs>=2.2.3
types-pytz>=2025.2

# Added pandas type stubs
# Added dependency for pandas-stubs 